+++ openstf/model/serializer.py
@@ -21,7 +21,7 @@ from mlflow.tracking import MlflowClient
 from openstf_dbc.services.prediction_job import PredictionJobDataClass
 from plotly import graph_objects
 
-from openstf.dataclasses.model_specifications import ModelSpecificationDataClass
+from openstf.data_classes.model_specifications import ModelSpecificationDataClass
 from openstf.metrics.reporter import Report
 from openstf.model.regressors.regressor import OpenstfRegressor
