+++ openstf/pipeline/optimize_hyperparameters.py
@@ -9,7 +9,7 @@ import pandas as pd
 import structlog
 from openstf_dbc.services.prediction_job import PredictionJobDataClass
 
-from openstf.dataclasses.model_specifications import ModelSpecificationDataClass
+from openstf.data_classes.model_specifications import ModelSpecificationDataClass
 from openstf.exceptions import (
     InputDataInsufficientError,
     InputDataWrongColumnOrderError,
